% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgi_batched_fetch.R
\name{batched_fetch}
\alias{batched_fetch}
\title{Batched Data Fetching Functions}
\description{
Internal functions that add batching capabilities to existing
date-based VGI functions to reduce API stress.
}
\keyword{internal}
