% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgi_rate_limiter.R
\name{should_use_rate_limiting}
\alias{should_use_rate_limiting}
\title{Check if Rate Limiting Should Be Applied}
\usage{
should_use_rate_limiting(num_dates, num_metrics = 1)
}
\arguments{
\item{num_dates}{Number of dates to process}

\item{num_metrics}{Number of metrics being fetched}
}
\value{
Logical indicating if rate limiting is recommended
}
\description{
Determines if rate limiting should be applied based on the
expected number of API calls.
}
