% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgi_cache_system.R
\name{vgi_cache_stats}
\alias{vgi_cache_stats}
\title{Get Cache Statistics}
\usage{
vgi_cache_stats()
}
\value{
Invisible list with cache statistics
}
\description{
Display information about the current cache.
}
