% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgi_insights_units.R
\name{format_date}
\alias{format_date}
\title{Format date for API}
\usage{
format_date(date)
}
\arguments{
\item{date}{Date or character string to format}
}
\value{
Character string in YYYY-MM-DD format
}
\description{
Format date for API
}
\keyword{internal}
