% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgi_game_summary_yoy.R
\name{print.vgi_yoy_comparison}
\alias{print.vgi_yoy_comparison}
\title{Print Method for Year-over-Year Comparison}
\usage{
\method{print}{vgi_yoy_comparison}(x, ...)
}
\description{
Print Method for Year-over-Year Comparison
}
